refactor(instant): Make instant an optional dependency
The `instant` crate is used by `backoff` to support the `Instant` type
when compiling to WASM, but it is currently unmaintained. This PR makes
the dependency optional, so that clients that do not need this feature
can avoid taking the dependency.

The feature is enabled by default so this change should not be breaking,
but it allows downstream packages to disable default features if they
don't need to take the dependency.

A separate PR is required to change the dependency to something like
`web-time` that is currently maintained.

## Test plan

Ran tests with and without the feature enabled:

```
cargo nextest run
cargo nextest run --no-default-features
cargo nextest run --no-default-features --features wasm-bindgen
cargo nextest run --features wasm-bindgen
```